Add methods for changing hash table size
Add test function stubs
Use uint64 across all attributes

+++ pkg/helper/hasher.go
@@ -1,5 +1,5 @@
 package helper
 
-func HashKey(key []byte, buckets uint32) uint32 {
-	return uint32(key[0]) % buckets
+func HashKey(key []byte, buckets uint64) uint64 {
+	return uint64(key[0]) % buckets
 }
